Add Module Courses

+++ modules/Courses/routes/web.php
@@ -3,17 +3,16 @@
 use Illuminate\Support\Facades\Route;
 use Modules\Courses\src\Http\Controllers\CourseController;
 
-Route::group(['prefix' => 'admin', 'middleware' => 'web', 'name' => 'admin.'], function () {
-    Route::group(['prefix' => 'courses', 'name' => 'courses.'], function () {
-        Route::get('/', [CourseController::class, 'index'])->name('index');
-        Route::get('/data', [CourseController::class, 'data'])->name('data');
+Route::group(['prefix' => 'admin', 'middleware' => 'web'], function () {
+    Route::group(['prefix' => 'courses'], function () {
+        Route::get('/', [CourseController::class, 'index'])->name('admin.courses.index');
 
-        Route::get('/create', [CourseController::class, 'create'])->name('create');
+        Route::get('/create', [CourseController::class, 'create'])->name('admin.courses.create');
 
-        Route::post('/store', [CourseController::class, 'store'])->name('store');
-        Route::get('/edit/{course}', [CourseController::class, 'edit'])->name('edit');
-        Route::put('/update/{course}', [CourseController::class, 'update'])->name('update');
-        Route::delete('/delete/{course}', [CourseController::class, 'delete'])->name('delete');
+        Route::post('/store', [CourseController::class, 'store'])->name('admin.courses.store');
+        Route::get('/edit/{course}', [CourseController::class, 'edit'])->name('admin.courses.edit');
+        Route::put('/update/{course}', [CourseController::class, 'update'])->name('admin.courses.update');
+        Route::delete('/delete/{course}', [CourseController::class, 'delete'])->name('admin.courses.delete');
 
     });
 });
